Fix of previous observations
it has been made one query to get the produts that are going out and other query to get the products that are go in in the 'susitutive' category. Also, if the id_category for the backup ( from table atributos_producto) doesn´t get updated, the product's ref_id it will be removed from the list of products that are going into the 'sustitutive' category. Then concatenate both results using the format used to change of category.